Only set TestXml when relevant
TestXml is now set when:
* We change from the tab "Properties" to the tab "XML"
* Or when the tab "XML" is visible and the selected item in the tree changes

This improves the performance of most operations in the gui
(i.e. all operations as long as the "XML" tab is not selected). The only
downside is that going back and forth between the tabs will retrigger
a call to `DisplayXml`.

To improve the experience even more `GetFullXml` and `XmlHelper.ToRtfString`
could be done asynchronous and results could be cached.

Relates to #241